// frontend/taro-app/packages/core/src/rpc/common/v1/types.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               unknown
// source: common/v1/types.proto

/* eslint-disable */

/** double 类型数组包装消息。 */
export interface DoubleValues {
  /** 值列表。 */
  value: number[];
}

/** float 类型数组包装消息。 */
export interface FloatValues {
  /** 值列表。 */
  value: number[];
}

/** int64 类型数组包装消息。 */
export interface Int64Values {
  /** 值列表。 */
  value: number[];
}

/** uint64 类型数组包装消息。 */
export interface UInt64Values {
  /** 值列表。 */
  value: number[];
}

/** int32 类型数组包装消息。 */
export interface Int32Values {
  /** 值列表。 */
  value: number[];
}

/** uint32 类型数组包装消息。 */
export interface UInt32Values {
  /** 值列表。 */
  value: number[];
}

/** bool 类型数组包装消息。 */
export interface BoolValues {
  /** 值列表。 */
  value: boolean[];
}

/** string 类型数组包装消息。 */
export interface StringValues {
  /** 值列表。 */
  value: string[];
}

/** bytes 类型数组包装消息。 */
export interface BytesValues {
  /** 值列表。 */
  value: Uint8Array[];
}

/** 密码密文。 */
export interface PasswordCrypto {
  /** 临时密钥ID */
  key_id: string;
  /** 随机值 */
  nonce: string;
  /** 加密算法 */
  algorithm: string;
  /** 加密后的对称密钥 */
  encrypted_key: string;
  /** 初始化向量 */
  iv: string;
  /** 密码加密数据 */
  ciphertext: string;
}